+++ backend/canisters/community/impl/src/lifecycle/post_upgrade.rs
@@ -1,14 +1,17 @@
 use crate::jobs::import_groups::finalize_group_import;
 use crate::lifecycle::{init_env, init_state};
 use crate::memory::{get_stable_memory_map_memory, get_upgrades_memory};
-use crate::{Data, read_state};
+use crate::model::events::CommunityEventInternal;
+use crate::{Data, mutate_state, read_state};
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
+use chat_events::{ChatEventInternal, EventKey, Reader};
 use community_canister::post_upgrade::Args;
 use ic_cdk::post_upgrade;
 use instruction_counts_log::InstructionCountFunctionId;
 use stable_memory::get_reader;
 use tracing::info;
+use types::{ChannelCreated, EventIndex, EventWrapperInternal};
 
 #[post_upgrade]
 #[trace]
@@ -40,5 +43,57 @@ fn post_upgrade(args: Args) {
         state
             .data
             .record_instructions_count(InstructionCountFunctionId::PostUpgrade, now);
+    });
+
+    // TODO: Delete after communities are upgraded
+    mutate_state(|state| {
+        // Read all the existing community events onto the heap
+        let mut community_events = state.data.events.read_all_events();
+
+        // Check that all events have an event index corresponding to their index in the vector
+        for (index, event) in community_events.iter().enumerate() {
+            if event.index != EventIndex::from(index as u32) {
+                panic!("Event index mismatch: expected {}, got {}", index, event.index);
+            }
+        }
+
+        // Append ChannelCreated events to CommunityEvents
+        for channel in state.data.channels.iter() {
+            if let Some(wrapper) = channel
+                .chat
+                .events
+                .main_events_reader()
+                .get_event(EventKey::EventIndex(EventIndex::from(0)))
+            {
+                if let ChatEventInternal::GroupChatCreated(event) = wrapper.event {
+                    let new_event_wrapper = EventWrapperInternal {
+                        index: EventIndex::from(0),
+                        timestamp: channel.date_imported.unwrap_or(channel.chat.date_created),
+                        expires_at: None,
+                        event: CommunityEventInternal::ChannelCreated(Box::new(ChannelCreated {
+                            channel_id: channel.id,
+                            is_public: channel.chat.is_public.value,
+                            name: event.name.clone(),
+                            created_by: event.created_by,
+                        })),
+                    };
+
+                    community_events.push(new_event_wrapper);
+                } else {
+                    panic!("Expected GroupChatCreated event, found {:?}", wrapper.event);
+                }
+            }
+        }
+
+        // Stable sort the community events by timestamp - events with the same timestamp will stay in the same order
+        community_events.sort_by_key(|event| event.timestamp);
+
+        // Iterate through the community events and rewrite the event indexes to match the vector index
+        for (index, event) in community_events.iter_mut().enumerate() {
+            event.index = EventIndex::from(index as u32);
+        }
+
+        // Write the updated community events back to stable memory
+        state.data.events.overwrite_all_events(community_events);
     });
 }