+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Introduce and backdate a ChannelCreated event ([#8234](https://github.com/open-chat-labs/open-chat/pull/8234))
+
 ## [[2.0.1802](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1802-local_user_index)] - 2025-06-26
 
 ### Added